# PMS.swagger.json
{
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/pet-system/ai/session/create": {
      "post": {
        "summary": "createSession",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": "",
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultString"
            }
          }
        },
        "security": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/pet-system/ai/chat": {
      "post": {
        "summary": "chat",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AIAgentRequest",
              "description": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": {
                  "messageId": "",
                  "sessionId": "",
                  "content": "",
                  "answer": "",
                  "toolCalls": [
                    {
                      "id": "",
                      "type": "",
                      "name": "",
                      "arguments": "",
                      "result": ""
                    }
                  ],
                  "needHuman": false,
                  "suggestions": [
                    ""
                  ],
                  "tokensUsed": 0,
                  "latencyMs": 0
                },
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultAIAgentResponse"
            }
          }
        },
        "security": [],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ]
      }
    },
    "/pet-system/ai/history/{sessionId}": {
      "get": {
        "summary": "getHistory",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": [
                  {
                    "id": 0,
                    "sessionId": "",
                    "messageId": "",
                    "role": "",
                    "content": "",
                    "toolCalls": "",
                    "toolCallId": "",
                    "tokensUsed": 0,
                    "latencyMs": 0,
                    "feedback": 0,
                    "createdAt": ""
                  }
                ],
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultListAiChatMessage"
            }
          }
        },
        "security": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/pet-system/ai/memory/{sessionId}": {
      "delete": {
        "summary": "clearMemory",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": "",
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultString"
            }
          }
        },
        "security": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/pet-system/ai/transfer": {
      "post": {
        "summary": "transferToHuman",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "reason",
            "in": "query",
            "description": "",
            "required": false,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": "",
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultString"
            }
          }
        },
        "security": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/pet-system/ai/feedback": {
      "post": {
        "summary": "submitFeedback",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "messageId",
            "in": "query",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "score",
            "in": "query",
            "description": "",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": "",
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultString"
            }
          }
        },
        "security": [],
        "produces": [
          "application/json"
        ]
      }
    },
    "/pet-system/ai/suggestions": {
      "get": {
        "summary": "getSuggestions",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "default": "Bearer {{ACCESS_TOKEN}}",
            "type": "string",
            "x-example": "Bearer {{ACCESS_TOKEN}}"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "examples": {
              "application/json": {
                "code": 0,
                "message": "",
                "data": [
                  ""
                ],
                "exception": ""
              }
            },
            "schema": {
              "$ref": "#/definitions/CommonResultListString"
            }
          }
        },
        "security": [],
        "produces": [
          "application/json"
        ]
      }
    },
...
  },
  "security": [],
  "swagger": "2.0",
  "definitions": {
...
    "CommonResultString": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "错误码"
        },
        "message": {
          "type": "string",
          "description": "错误信息"
        },
        "data": {
          "type": "string",
          "description": "数据结果"
        },
        "exception": {
          "type": "string",
          "description": "异常信息"
        }
      }
    },
...
    "ToolCall": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "description": ""
        },
        "type": {
          "type": "string",
          "description": ""
        },
        "name": {
          "type": "string",
          "description": ""
        },
        "arguments": {
          "type": "string",
          "description": ""
        },
        "result": {
          "type": "string",
          "description": ""
        }
      }
    },
...
    "AIAgentResponse": {
      "type": "object",
      "properties": {
        "messageId": {
          "type": "string",
          "description": ""
        },
        "sessionId": {
          "type": "string",
          "description": ""
        },
        "content": {
          "type": "string",
          "description": ""
        },
        "answer": {
          "type": "string",
          "description": ""
        },
        "toolCalls": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ToolCall",
            "description": "com.hongjie.pms.AI.modules.dto.ToolCall"
          },
          "description": ""
        },
        "needHuman": {
          "type": "boolean",
          "description": ""
        },
        "suggestions": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": ""
        },
        "tokensUsed": {
          "type": "integer",
          "description": ""
        },
        "latencyMs": {
          "type": "integer",
          "description": ""
        }
      }
    },
...
    "CommonResultAIAgentResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "错误码"
        },
        "message": {
          "type": "string",
          "description": "错误信息"
        },
        "data": {
          "$ref": "#/definitions/AIAgentResponse",
          "description": "数据结果"
        },
        "exception": {
          "type": "string",
          "description": "异常信息"
        }
      }
    },
...
    "ChatHistory": {
      "type": "object",
      "properties": {
        "role": {
          "type": "string",
          "description": ""
        },
        "content": {
          "type": "string",
          "description": ""
        }
      }
    },
...
    "AIAgentRequest": {
      "type": "object",
      "properties": {
        "sessionId": {
          "type": "string",
          "description": ""
        },
        "message": {
          "type": "string",
          "description": ""
        },
        "userId": {
          "type": "integer",
          "description": "",
          "format": "int64"
        },
        "stream": {
          "type": "boolean",
          "description": "",
          "default": false
        },
        "history": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ChatHistory",
            "description": "com.hongjie.pms.AI.modules.dto.request.AIAgentRequest.ChatHistory"
          },
          "description": ""
        }
      },
      "required": [
        "sessionId",
        "message"
      ]
    },
...
    "AiChatMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "",
          "format": "int64"
        },
        "sessionId": {
          "type": "string",
          "description": ""
        },
        "messageId": {
          "type": "string",
          "description": ""
        },
        "role": {
          "type": "string",
          "description": ""
        },
        "content": {
          "type": "string",
          "description": ""
        },
        "toolCalls": {
          "type": "string",
          "description": ""
        },
        "toolCallId": {
          "type": "string",
          "description": ""
        },
        "tokensUsed": {
          "type": "integer",
          "description": ""
        },
        "latencyMs": {
          "type": "integer",
          "description": ""
        },
        "feedback": {
          "type": "integer",
          "description": ""
        },
        "createdAt": {
          "type": "string",
          "description": ""
        }
      }
    },
...
    "CommonResultListAiChatMessage": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "错误码"
        },
        "message": {
          "type": "string",
          "description": "错误信息"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AiChatMessage",
            "description": "com.hongjie.pms.AI.modules.entity.AiChatMessage"
          },
          "description": "数据结果"
        },
        "exception": {
          "type": "string",
          "description": "异常信息"
        }
      }
    },
...
    "CommonResultListString": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "description": "错误码"
        },
        "message": {
          "type": "string",
          "description": "错误信息"
        },
        "data": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": "数据结果"
        },
        "exception": {
          "type": "string",
          "description": "异常信息"
        }
      }
    },
...
  },
  "securityDefinitions": {},
...
}